Add find_rhyme_pairs method to Poem
The goal of this method is to find whether a given poem exhibits a particular
rhyme pair; the main use case is to try it with surprising pairs (things that
supposedly should not rhyme, but do). The return value is a set of all pairs
found, so that an empty set means the poem does not exhibit the searched
pattern.