Add align_communities function (#22)
When presented with two corpora (or divisions of a single corpora) and given
their communities of rhymes, it is interesting to align the communities so that
we can later compare e.g. the composition of the 歌/戈 community of corpus A
with that of corpus B, and see if there are merges on one side or the other.