Add filter_by_period function for CBDB (#26)
This allows restricting the people returned by get_person to a specific time
range; if birth and death years are known, these are used; otherwise, they are
estimated from the other by subtracting / adding 80 years (i.e. if we only know
a person to be born in 880AD, we estimate their death in 960AD). If none of
these is known, the Index year is used.