Use @flaky for flaky tests
`shuffle`, by its nature, can return the same graph as the input, making the
test that it's different fail. This can be mitigated by having a much larger
graph (in which case the probability is very low), but it seems more robust
to mark the test as flaky and have it retried.